# docs/openapi.yaml
openapi: 3.1.0
info:
  title: GuardianWAF API
  description: GuardianWAF REST API for security monitoring, configuration, and management.
  version: 1.0.0
  contact:
    name: GuardianWAF
    url: https://github.com/guardianwaf/guardianwaf
  license:
    name: MIT
servers:
  - url: https://localhost:9443
    description: Default dashboard endpoint
security:
  - ApiKeyAuth: []
  - SessionAuth: []
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    SessionAuth:
      type: apiKey
      in: cookie
      name: gwaf_session
    AdminKeyAuth:
      type: apiKey
      in: header
      name: X-Admin-Key
paths:
...
  /api/v1/analytics/traffic:
    get:
      tags: [Analytics]
      summary: Get traffic statistics
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Traffic statistics
  /api/v1/analytics/trends:
    get:
      tags: [Analytics]
      summary: Get trend analysis
      parameters:
        - name: metric
          in: query
          schema:
            type: string
        - name: interval
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Trend data
  /api/v1/analytics/geo:
    get:
      tags: [Analytics]
      summary: Get geographic distribution
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Geo distribution data
  /api/v1/analytics/comparison:
    get:
      tags: [Analytics]
      summary: Compare two time periods
      parameters:
        - name: current_from
          in: query
          schema:
            type: string
            format: date-time
        - name: current_to
          in: query
          schema:
            type: string
            format: date-time
        - name: previous_from
          in: query
          schema:
            type: string
            format: date-time
        - name: previous_to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Period comparison data
...
  /api/v1/analytics/timeseries:
    get:
      tags: [Analytics]
      summary: Get time series data
      parameters:
        - name: name
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Time series data

  # ── Multi-Tenant Admin ────────────────────────────────
...